factory module definition
